Development

+++ app/Controllers/Http/SkillController.js
@@ -20,8 +20,8 @@ class SkillController {
    * @param {View} ctx.view
    */
   async index ({ request, response, view }) {
-    const { master } = request
-    const skills = await master.skills().fetch()
+    const { adventure } = request
+    const skills = await Skill.query().with('icon').where('adventure_id', adventure.id).fetch()
     return skills
   }
 
@@ -34,9 +34,9 @@ class SkillController {
    * @param {Response} ctx.response
    */
   async store ({ request, response }) {
-    const { master } = request
-    const data = request.only(['name', 'description', 'type', 'value'])
-    const skill = await Skill.create({ ...data, master_id: master.id })
+    const { master, adventure_id } = request
+    const data = request.only(['name', 'description', 'type', 'value', 'mana_cost', 'icon_id'])
+    const skill = await Skill.create({ ...data, adventure_id, master_id: master.id })
     return skill
   }
 
@@ -61,7 +61,20 @@ class SkillController {
    * @param {Response} ctx.response
    */
   async update ({ params, request, response }) {
-
+    // const { master, adventure_id } = request
+    const data = request.only(['name', 'description', 'type', 'value', 'mana_cost', 'icon_id'])
+    const skill = await Skill.find(params.id)
+    if (!skill) {
+      return response.status(404).json({
+        error: {
+          message: 'skill not found'
+        }
+      })
+    }
+    await skill.merge(data)
+    await skill.save()
+    await skill.load('icon')
+    return skill
   }
 
   /**
@@ -74,11 +87,11 @@ class SkillController {
    */
   async destroy ({ params, request, response }) {
     const skill = await Skill.find(params.id)
-    await skill.delete()
+    if (skill) {
+      await skill.delete()
+    }
     return {
-      ok: {
-        message: 'Skill deleted with sucess'
-      }
+      ok: true
     }
   }
 }
